Extract builder to separate class and expose through static constructor on interface.
