+++ library/src/main/java/com/novoda/storagepathfinder/DeviceStorageInspector.java
@@ -36,61 +36,8 @@ public interface DeviceStorageInspector {
      */
     List<StoragePath> getSecondaryStorageApplicationPath();
 
-    static Builder builder(Context context) {
-        return Builder.newInstance(context);
-    }
-
-    final class Builder {
-
-        private final Context context;
-        private FileSystem fileSystem;
-        private DeviceFeatures deviceFeatures;
-        private AndroidSystem androidSystem;
-
-        private static Builder newInstance(Context context) {
-            Context applicationContext = context.getApplicationContext();
-            ExternalStorageDirectories externalStorageDirectories = new AndroidExternalStorageDirectories(applicationContext);
-            FileSystem fileSystem = new AndroidFileSystem(externalStorageDirectories);
-            DeviceFeatures deviceFeatures = new AndroidDeviceFeatures();
-            AndroidSystem androidSystem = new AndroidSystem();
-            return new Builder(
-                    applicationContext,
-                    fileSystem,
-                    deviceFeatures,
-                    androidSystem
-            );
-        }
-
-        private Builder(Context context, FileSystem fileSystem, DeviceFeatures deviceFeatures, AndroidSystem androidSystem) {
-            this.context = context;
-            this.fileSystem = fileSystem;
-            this.deviceFeatures = deviceFeatures;
-            this.androidSystem = androidSystem;
-        }
-
-        public Builder withFileSystem(FileSystem fileSystem) {
-            this.fileSystem = fileSystem;
-            return this;
-        }
-
-        public Builder withDeviceFeatures(DeviceFeatures deviceFeatures) {
-            this.deviceFeatures = deviceFeatures;
-            return this;
-        }
-
-        public Builder withAndroidSystem(AndroidSystem androidSystem) {
-            this.androidSystem = androidSystem;
-            return this;
-        }
-
-        public DeviceStorageInspector build() {
-            return new AndroidDeviceStorageInspector(
-                    context,
-                    fileSystem,
-                    deviceFeatures,
-                    androidSystem
-            );
-        }
+    static DeviceStorageInspectorBuilder builder(Context context) {
+        return DeviceStorageInspectorBuilder.newInstance(context);
     }
 
 }
